# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Management Chatbot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: plan
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Register a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /members/{id}/progress:
    get:
      operationId: getMemberProgress
      tags: [members]
      summary: Get member progress tracking entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressEntry"
    post:
      operationId: addMemberProgress
      tags: [members]
      summary: Add a progress entry for member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressEntryInput"
      responses:
        "201":
          description: Progress entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressEntry"
  /members/{id}/renew:
    post:
      operationId: renewMembership
      tags: [members]
      summary: Renew a member's membership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewalInput"
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # ── ATTENDANCE ────────────────────────────────────────────────────────────
...
  /trainers:
    get:
      operationId: listTrainers
      tags: [trainers]
      summary: List all trainers
      responses:
        "200":
          description: Trainers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainer"
    post:
      operationId: createTrainer
      tags: [trainers]
      summary: Add a new trainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerInput"
      responses:
        "201":
          description: Created trainer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
  /trainers/{id}:
    get:
      operationId: getTrainer
      tags: [trainers]
      summary: Get trainer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trainer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
    patch:
      operationId: updateTrainer
      tags: [trainers]
      summary: Update trainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerUpdate"
      responses:
        "200":
          description: Updated trainer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
    delete:
      operationId: deleteTrainer
      tags: [trainers]
      summary: Delete a trainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trainers/{id}/members:
    get:
      operationId: getTrainerMembers
      tags: [trainers]
      summary: Get members assigned to a trainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assigned members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /trainers/{id}/bookings:
    get:
      operationId: getTrainerBookings
      tags: [trainers]
      summary: Get trainer bookings/schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trainer bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainerBooking"
  /trainers/bookings:
    post:
      operationId: createTrainerBooking
      tags: [trainers]
      summary: Book a trainer session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerBooking"

  # ── DIET PLANS ────────────────────────────────────────────────────────────
  /diet-plans:
    get:
      operationId: listDietPlans
      tags: [plans]
      summary: List diet plans
      parameters:
        - name: goal
          in: query
          schema:
            type: string
        - name: trainerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Diet plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DietPlan"
    post:
      operationId: createDietPlan
      tags: [plans]
      summary: Create a diet plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DietPlanInput"
      responses:
        "201":
          description: Diet plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DietPlan"
  /workout-plans:
    get:
      operationId: listWorkoutPlans
      tags: [plans]
      summary: List workout plans
      parameters:
        - name: goal
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Workout plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlan"
    post:
      operationId: createWorkoutPlan
      tags: [plans]
      summary: Create a workout plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutPlanInput"
      responses:
        "201":
          description: Workout plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"

  # ── LEADS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── CHAT SCHEMAS ────────────────────────────────────────────────────────
...
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        gender:
          type: string
        age:
          type: integer
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        goal:
          type: ["string", "null"]
        plan:
          type: string
        joiningDate:
          type: string
        expiryDate:
          type: string
        paymentStatus:
          type: string
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        bmi:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - gender
        - age
        - plan
        - joiningDate
        - expiryDate
        - paymentStatus
        - status
        - createdAt
    MemberInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        gender:
          type: string
        age:
          type: integer
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        goal:
          type: ["string", "null"]
        plan:
          type: string
        joiningDate:
          type: string
        paymentStatus:
          type: string
        trainerId:
          type: ["integer", "null"]
      required:
        - name
        - phone
        - gender
        - age
        - plan
        - joiningDate
        - paymentStatus
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        gender:
          type: string
        age:
          type: integer
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        goal:
          type: ["string", "null"]
        plan:
          type: string
        paymentStatus:
          type: string
        trainerId:
          type: ["integer", "null"]
        status:
          type: string
    RenewalInput:
      type: object
      properties:
        plan:
          type: string
        paymentStatus:
          type: string
      required:
        - plan
        - paymentStatus
    ProgressEntry:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        weight:
          type: ["number", "null"]
        bmi:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
      required:
        - id
        - memberId
        - recordedAt
    ProgressEntryInput:
      type: object
      properties:
        weight:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required: []

    # ── ATTENDANCE SCHEMAS ──────────────────────────────────────────────────
...
    Trainer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - memberCount
        - createdAt
    TrainerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
      required:
        - name
        - phone
    TrainerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
    TrainerBooking:
      type: object
      properties:
        id:
          type: integer
        trainerId:
          type: integer
        trainerName:
          type: ["string", "null"]
        memberId:
          type: integer
        memberName:
          type: ["string", "null"]
        scheduledAt:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - trainerId
        - memberId
        - scheduledAt
        - status
        - createdAt
    TrainerBookingInput:
      type: object
      properties:
        trainerId:
          type: integer
        memberId:
          type: integer
        scheduledAt:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - trainerId
        - memberId
        - scheduledAt

    # ── PLAN SCHEMAS ────────────────────────────────────────────────────────
    DietPlan:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        goal:
          type: string
        content:
          type: string
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - goal
        - content
        - createdAt
    DietPlanInput:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        content:
          type: string
        trainerId:
          type: ["integer", "null"]
      required:
        - title
        - goal
        - content
    WorkoutPlan:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        goal:
          type: string
        content:
          type: string
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - goal
        - content
        - createdAt
    WorkoutPlanInput:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        content:
          type: string
        trainerId:
          type: ["integer", "null"]
      required:
        - title
        - goal
        - content

    # ── LEAD SCHEMAS ────────────────────────────────────────────────────────
...