# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Management Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  gender?: string;
  age?: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  goal?: string | null;
  plan?: string;
  paymentStatus?: string;
  /** @nullable */
  trainerId?: number | null;
  status?: string;
}