# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Management Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  id: number;
  memberId: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  notes?: string | null;
  recordedAt: string;
}