# lib/api-zod/src/generated/types/trainerBooking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Management Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface TrainerBooking {
  id: number;
  trainerId: number;
  /** @nullable */
  trainerName?: string | null;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  scheduledAt: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}